[ALPS02456698] aee: fix Coverity Scanned Code Defect
[Detail]
Fix the Unintentional integer overflow issue

[Solution]

MTK-Commit-Id: c401f0553bc401b5e64adeabba64b5e7bb583b1d

CR-Id: ALPS02456698
Feature: Others

+++ drivers/misc/mediatek/aee/ipanic/ipanic_mtd.c
@@ -74,7 +74,7 @@ static int ipanic_mtd_block_read(struct ipanic_mtd_data *ctx, off_t file_offset,
 		page_no = file_offset / ctx->mtd->writesize;
 		page_offset = file_offset % ctx->mtd->writesize;
 
-		rc = ipanic_mtd_block_read_single(ctx, page_no * ctx->mtd->writesize, buf);
+		rc = ipanic_mtd_block_read_single(ctx, page_no * (loff_t)ctx->mtd->writesize, buf);
 		if (rc < 0) {
 			LOGE("mtd read error page_no(%d) error(%d)\n", page_no, rc);
 			return -1;
